Prototype
Added prototype scene playing song, with a few spawn points with random timers throwing blocks towards player camera and gun. No interaction allowing gun to destroy blocks yet, or code to sync with song.

Added incomplete Menu scene (I stashed stuff for later).

I might've broken some materials, everything loaded pink for me :P

+++ CGP603/Assets/TargetMove.cs
@@ -6,7 +6,7 @@ public class TargetMove : MonoBehaviour
 {
 
     public GameObject Target;
-    public int Speed;
+    public float Speed;
     // Start is called before the first frame update
     void Start()
     {
@@ -24,7 +24,6 @@ public class TargetMove : MonoBehaviour
         if(collision.gameObject.tag == "Wall")
         {
             Destroy(Target);
-
         }
     }
 }
